:recycle: refactor: MemberJpaRepo와 MemberRepo 테스트코드 통합
