+++ src/test/java/com/springles/repository/MemberJpaRepositoryTest.java
@@ -1,8 +1,11 @@
 package com.springles.repository;
 
+import com.springles.domain.dto.member.MemberCreateRequest;
 import com.springles.domain.entity.Member;
 import com.springles.exception.CustomException;
+import com.springles.service.MemberService;
 import jakarta.transaction.Transactional;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +15,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import java.util.List;
 import java.util.Optional;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
@@ -22,7 +26,41 @@ class MemberJpaRepositoryTest {
     MemberJpaRepository memberJpaRepository;
 
     @Autowired
+    MemberService memberService;
+
+    @Autowired
     PasswordEncoder passwordEncoder;
+
+    @AfterEach
+    public void deleteAll(){
+        memberJpaRepository.deleteAll();
+    }
+
+    @Test
+    @DisplayName("멤버 이름 조회 테스트 ( 대 소문자 구분 하지 않음 )")
+    public void findByOpenTrueTest() {
+
+        // given
+        for (int i = 1; i <= 2; i++) {
+            memberService.signUp(MemberCreateRequest.builder()
+                    .memberName("testUser" + i)
+                    .password("1")
+                    .passwordConfirm("1")
+                    .email("1@")
+                    .role("user")
+                    .isDeleted(false).build());
+        }
+
+        //when
+        Optional<List<Member>> findByNameIgnoreCase = memberJpaRepository.findAllByMemberNameContainingIgnoreCase("uSeR");
+
+        //then
+        List<Member> members = findByNameIgnoreCase.get();
+        for (int i = 0, j = 1; i < members.size(); i++, j++) {
+            assertThat(members.get(i).getMemberName()).isEqualTo("testUser" + j);
+        }
+    }
+
 
     @Test
     @DisplayName("memberName에 해당하는 Member 존재여부 확인 테스트")
@@ -42,6 +80,7 @@ class MemberJpaRepositoryTest {
         // then
         assertTrue(result);
     }
+
 
     @Test
     @DisplayName("memberName에 해당하는 Member 호출 테스트")
@@ -67,6 +106,7 @@ class MemberJpaRepositoryTest {
         assertEquals(optionalMember.get().getEmail(), "user1@gmail.com");
     }
 
+
     @Test
     @DisplayName("memberName에 해당하는 Member 삭제 테스트(삭제: isDeleted == true)")
     void deleteByMemberName() {
@@ -86,6 +126,7 @@ class MemberJpaRepositoryTest {
         // then
         assertTrue(optionalMember.get().getIsDeleted());
     }
+
 
     @Test
     @DisplayName("email에 해당하는 Member 전체 호출 테스트")
@@ -113,6 +154,7 @@ class MemberJpaRepositoryTest {
         // then
         assertEquals(memberList.size(), 2);
     }
+
 
     @Test
     @DisplayName("email과 memberName에 해당하는 Member 호출 테스트")
